+++ lib/src/ui/anchor.dart
@@ -7,7 +7,7 @@ part of dart_web_toolkit_ui;
  * A widget that represents a simple [:a:] element.
  */
 class Anchor extends FocusWidget implements HasHorizontalAlignment,
-  HasName, HasHtml, HasWordWrap, //HasDirection,
+  HasName, HasHtml, HasWordWrap,
   HasDirectionEstimator, HasDirectionalSafeHtml {
 
   /**
@@ -18,6 +18,41 @@ class Anchor extends FocusWidget implements HasHorizontalAlignment,
   static String DEFAULT_HREF = "javascript:;";
 
   HorizontalAlignmentConstant _horzAlign;
+  
+  DirectionalTextHelper _directionalTextHelper;
+  
+  /**
+   * Creates an Anchor widget that wraps an existing &lt;a&gt; element.
+   * 
+   * This element must already be attached to the document. If the element is
+   * removed from the document, you must call
+   * {@link RootPanel#detachNow(Widget)}.
+   * 
+   * @param element the element to be wrapped
+   */
+  factory Anchor.wrap(dart_html.Element element) {
+    // Assert that the element is attached.
+    //assert Document.get().getBody().isOrHasChild(element);
+
+    Anchor anchor = new Anchor.fromElement(element);
+
+    // Mark it attached and remember it for cleanup.
+    anchor.onAttach();
+    RootPanel.detachOnWindowClose(anchor);
+
+    return anchor;
+  }
+  
+  /**
+   * This constructor may be used by subclasses to explicitly use an existing
+   * element. This element must be an &lt;a&gt; element.
+   * 
+   * @param element the element to be used
+   */
+  Anchor.fromElement(dart_html.AnchorElement element) {
+    setElement(element);
+    _directionalTextHelper = new DirectionalTextHelper(getAnchorElement(), /* is inline */ true);
+  }
 
   /**
    * Creates an empty anchor.
@@ -28,9 +63,14 @@ class Anchor extends FocusWidget implements HasHorizontalAlignment,
   Anchor([bool useDefaultHref = false]) {
     setElement(new dart_html.AnchorElement());
     clearAndSetStyleName("dwt-Anchor");
+    _directionalTextHelper = new DirectionalTextHelper(getAnchorElement(), /* is inline */true);
     if (useDefaultHref) {
       href = DEFAULT_HREF;
     }
+  }
+  
+  DirectionEstimator getDirectionEstimator() {
+    return _directionalTextHelper.getDirectionEstimator();
   }
 
   /**
@@ -52,7 +92,21 @@ class Anchor extends FocusWidget implements HasHorizontalAlignment,
     _horzAlign = align;
     getElement().style.textAlign = align.getTextAlignString();
   }
+  
+  //***************************
+  // Implementation of HasFocus
+  //***************************
+  /**
+   * Gets the widget's position in the tab index.
+   *
+   * @return the widget's tab index
+   */
+  int get tabIndex => getAnchorElement().tabIndex;
 
+  void setTabIndex(int index) {
+    getAnchorElement().tabIndex = index;
+  }
+  
   //***********
   // PROPERTIES
   //***********
@@ -107,7 +161,7 @@ class Anchor extends FocusWidget implements HasHorizontalAlignment,
   /**
    * Get the anchor's wordWrap.
    */
-  bool get wordWrap => getAnchorElement().style.whiteSpace == WhiteSpace.NORMAL;
+  bool get wordWrap => getAnchorElement().style.whiteSpace == WhiteSpace.NORMAL.value;
 
   /**
    * Sets the anchor's html.
@@ -125,26 +179,71 @@ class Anchor extends FocusWidget implements HasHorizontalAlignment,
   /**
    * Sets the anchor's text.
    */
-  String get text => getAnchorElement().text;
+  String get text => _directionalTextHelper.getTextOrHtml(false); //getAnchorElement().text;
 
   /**
    * Get the anchor's text.
    */
   void set text(String value) {
     assert(value != null);
-    getAnchorElement().text = value;
+    _directionalTextHelper.setTextOrHtml(value, false);
+  }
+  
+  void setText(String text, Direction dir) {
+    _directionalTextHelper.setTextOrHtml(text, false, dir);
   }
 
+  Direction getTextDirection() {
+    return _directionalTextHelper.getTextDirection();
+  }
+  
   /**
    * Sets the anchor's text.
    */
-  String get direction => getAnchorElement().dir;
+//  String get direction => getAnchorElement().dir;
+//  Direction getDirection() {
+//    return BidiUtils.getDirectionOnElement(getElement());
+//  }
 
   /**
    * Get the anchor's text.
    */
-  void set direction(String value) {
-    assert(value != null);
-    getAnchorElement().dir = value;
+//  void set direction(String value) {
+//    assert(value != null);
+//    getAnchorElement().dir = value;
+//  }
+  
+  /**
+   * {@inheritDoc}
+   * <p>
+   * See note at {@link #setDirectionEstimator(DirectionEstimator)}.
+   */
+  void enableDirectionEstimator(bool enabled) {
+    _directionalTextHelper.enableDefaultDirectionEstimator(enabled);
+  }
+  
+  /**
+   * {@inheritDoc}
+   * <p>
+   * Note: DirectionEstimator should be set before the widget has any content;
+   * it's highly recommended to set it using a constructor. Reason: if the
+   * widget already has non-empty content, this will update its direction
+   * according to the new estimator's result. This may cause flicker, and thus
+   * should be avoided.
+   */
+  void setDirectionEstimator(DirectionEstimator directionEstimator) {
+    _directionalTextHelper.setDirectionEstimator(directionEstimator);
+  }
+  
+  void setAccessKey(int key) {
+    //getAnchorElement().accessKey(Character.toString(key));
+  }
+  
+  void setFocus(bool focused) {
+    if (focused) {
+      getAnchorElement().focus();
+    } else {
+      getAnchorElement().blur();
+    }
   }
 }